bpf: refactor ct maps into lib/{nat,conntrack_map}.h
We need to reuse the connection tracker from NAT as well, therefore
move the map definitions out of bpf_lxc. Also, take out the ct delete
helpers since we'll add NAT deletion logic in there as well.
